Varias mejoras del código.

+++ snapweb.sh
@@ -14,6 +14,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+## Agradecimientos:
+# Óscar García Amor
 #################### I N O T I F Y 
 #Aumento el número de ficheros a monitorizar (si es necesario)!!
 #sysctl -w fs.inotify.max_user_watches=524288
@@ -21,23 +23,45 @@
 #Miro si en el fichero /etc/sysctl.conf ya está el valor dado de alta, sino, lo doy.d
 #echo fs.inotify.max_user_watches=524288 >>/etc/sysctl.conf
 ####################################
+##Funciones
+installed(){ type -p "$1" 2>/dev/null >/dev/null;}
+fatal(){ echo ; echo "Error: $@" >&2;${E:+exit $E};}
+help () {
+echo "La sintaxis es:"
+echo "snapweb.sh [-d] /ruta/web"
+echo " -d --> Desactiva la monitorización del directorio pasado como parámetro."
+exit
+}
 
+clear
 #Mail destino donde se recibirá el informe de firma de snapweb
 mail_destino=wllop@esat.es
 #Variables y sanitizamos $1 para poder crear el fichero en /etc/incron.d/$1 sin las /
 filesan=$(echo $1|tr -d /)
 filetmp=/tmp/JACK$filesan
 #Comprobamos que está el módulo incron
-incrond -? 2>/dev/null
-if [ $? -ne 0 ];  then
- echo "Es necesario instalar el servicio incron en su sistema"
- echo s|apt-get install incron
- if [ $? -ne 0 ];then
-   echo "Se ha producido un problema al instalar el servicio incron"
-   echo "Consulte con su administrador :P"
-   exit
- fi
+if ! installed incrond ; then
+  echo "Es necesario disponder del servicio incron instalado en su sistema. "
+  if installed apt-get; then
+      read -n 1 -p "¿Quieres que lo instale automáticamente? (s/n):" var
+       if [ "$var" = "s" ]; then
+              echo s|apt-get install incron || E=1 fatal "El servicio incron no ha podido instalarse."
+       else
+              E=2 fatal "Intente ejecutar snapweb.sh una vez tenga instalado el servicio incron."
+          exit
+       fi
+  fi
 fi
+#Comprobamos que jack.sh estuviera en /usr/local/snapweb/snap_back, sino copiamos o bajamos con wget
+if [ ! -e /usr/local/snapweb/jack.sh ]; then
+  [ -e ./jack.sh ] && cp -f ./jack.sh /usr/local/snapweb 2>/dev/null >/dev/null && chmod a+x /usr/local/snapweb/jack.sh 2>/dev/null >/dev/null || E=3 fatal "Es necesario el fichero jack.sh para continuar."
+fi 
+
+#Damos de alta el fichero de configuración en /etc/snapweb.conf
+if [ ! -e /etc/snapweb.conf ]; then
+  [ -e ./snapweb.conf ] && cp -f ./snapweb.conf /etc/snapweb.conf || E=4 fatal "No ha sido posible obtener el fichero de configuracion snapweb.conf, inténtelo más tarde. Gracias."
+fi
+
 #Compruebo la opción -d
 if [ "$1" = "-d" ];then
   if [ ! -d $2 ] || [ "$2" = "" ]; then
@@ -55,26 +79,19 @@ if [ "$1" = "-d" ];then
           exit
         fi
       fi
-      echo "Ups: Algo ha fallado al intentar deshabilitar el directorio $2. Inténtelo otra vez!"
-      exit
+      E=5 fatal "Algo ha fallado al intentar deshabilitar el directorio $2. Inténtelo otra vez!"
   fi
- echo "El directorio $2 no está siendo monitorizado."
- exit
+ E=6 fatal "El directorio $2 no está siendo monitorizado."
 fi
-#Comprobamos parámetros.
+
+#Comprobamos ayuda.
 if [ ! -d $1 ] || [ "$1" == "" ] || [ "$1" == "-h" ];
 then
-  echo "La sintaxis es:"
-  echo "snapweb.sh [-d] /ruta/web"
-  echo " -d --> Desactiva la monitorización del directorio pasado como parámetro."
-  exit
+  help
 fi
 
 #Compruebo que el directorio de las snap exista
-if [ ! -d /usr/local/snapweb/snap_back ]; then
-   mkdir -p /usr/local/snapweb/snap_back
-   chmod 750 /usr/local/snapweb/snap_back
-fi
+[ ! -d /usr/local/snapweb/snap_back ] && mkdir -p -m 750 /usr/local/snapweb/snap_back 2>/dev/null
 
 if [ ! -d "/usr/local/snapweb/snap_back/$filesan" ]; #Preparamos screenshots
 then
@@ -83,50 +100,15 @@ then
    echo "${#ruta}">>/usr/local/snapweb/snap_back/$filesan/.ruta #Con esto convertiremos rutas absolutas en relativas a snap_back
    echo "Se ha creado una nueva firma del directorio $1"|mail -s "SNAPWEB: Nueva Firma" $mail_destino
 else
-   if [ -e /usr/local/snapweb/snap_back/$filesan.2 ]; then
-       rm -fr /usr/local/snapweb/snap_back/$filesan.2 
-   fi
+   [ -e /usr/local/snapweb/snap_back/$filesan.2 ] && rm -fr /usr/local/snapweb/snap_back/$filesan.2 
    ruta=$1
    mv -f /usr/local/snapweb/snap_back/$filesan /usr/local/snapweb/snap_back/$filesan.2
    cp -fpr $1 /usr/local/snapweb/snap_back/$filesan
    echo "${#ruta}">>/usr/local/snapweb/snap_back/$filesan/.ruta #Con esto convertiremos rutas absolutas en relativas a snap_back
    echo "Se ha creado una firma del directorio $1"|mail -s "SNAPWEB: Firma reemplazada" $mail_destino
 fi
-#Comprobamos que jack.sh estuviera en /usr/local/snapweb/snap_back, sino copiamos o bajamos con wget
-if [ ! -e /usr/local/snapweb/jack.sh ]; then
-	if [ -e ./jack.sh ]; then
-		cp -f ./jack.sh /usr/local/snapweb
-		chmod a+x /usr/local/snapweb/jack.sh
-	else
-		echo "Descargando...."
-		wget -nv -T 15 http://desa.webnet.es/snapweb/jack.sh -O /usr/local/snapweb/jack.sh >/dev/null 2>/dev/null
-		if [ ! -e /usr/local/snapweb/jack.sh ]; then
-			echo "Error: No ha sido posible obtener el fichero jack.sh. Inténtelo más tarde. Gracias."
-			exit
-		fi 
-		chmod a+x /usr/local/snapweb/jack.sh
-	fi
-fi 
 
-#Damos de alta el fichero en el directorio /etc/incron.d
-if [ ! -e /etc/snapweb.conf ];
-  then
-    if [ -e ./snapweb.conf ]; then
-    cp -f ./snapweb.conf /etc/snapweb.conf
-   else
-    echo "Descargando fichero de configuracion...."
-    wget -nv -T 15 http://desa.webnet.es/snapweb/snapweb.conf -O /etc/snapweb.conf >/dev/null 2>/dev/null
-   fi
-   if [ ! -e /etc/snapweb.conf ]; then
-      echo "Error: No ha sido posible obtener el fichero de configuracion snapweb.cfg. Inténtelo más tarde. Gracias."
-      exit
-  fi 
-fi
-
-if [ -e /etc/incron.d/$filesan ];
-then
-   rm -f /etc/incron.d/$filesan*
-fi
+[ -e /etc/incron.d/$filesan ] && rm -f /etc/incron.d/$filesan*
 #Excluyos los directorios indicados en la variable exclude_dir
 excl_dire=$(grep exclude_dir /etc/snapweb.conf|cut -d= -f2)
 if [ "$excl_dire" != "" ]; then
